Moves renderColumns function into Utility folder
