+++ client/src/components/TrendingVideos.jsx
@@ -3,36 +3,8 @@ import TrendingVideoCard from './TrendingVideoCard.jsx'
 import { Card, Divider, Grid, Header, Icon, Image, Menu, Segment } from 'semantic-ui-react';
 import { setPitchCategoryToTop, setPitchCategoryToTrending } from '../actions/pitchCategory'
 import { connect } from 'react-redux';
+import renderColumns from '../Utility/renderColumns.js'
 
-const renderColumns = (array, columnAmount) => {
-  let map = []
-  let leftOverIndex = null
-  for (var i = 0; i < array.length; i += columnAmount) {
-    let innerArray = []
-    for (var j = i; j < columnAmount; j++) {
-      let innerItem = array[j]
-      if (innerItem) { innerArray.push(innerItem) }
-    }
-    if (innerArray.length > 0) { map.push(innerArray) }
-    if (i + columnAmount >= array.length) {
-      leftOverIndex = i
-    }
-  }
-  if (leftOverIndex) {
-    map.push(array.slice(leftOverIndex))
-  }
-  return <Grid container padded columns={columnAmount}>
-    {map.map((col, i) => {
-      return <Grid.Row key={i}>
-        {col.map((pitch, j) => (
-          <Grid.Column key={j}>
-            <TrendingVideoCard pitch={pitch} />
-          </Grid.Column>
-      ))}
-    </Grid.Row>
-  })}
-  </Grid>
-}
 
 class TrendingVideos extends React.Component {
   constructor(props) {
